@@ -1,22 +1,31 @@
 using AspNetCore.FrameworkAdapter;
-using FrameworkSDK;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using Console.FrameworkAdapter.Constructing;
+ using FrameworkSDK.Constructing;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Epic.Server
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            new DefaultAppFactory(args)
-                .UseAspNetCore(CreateHostBuilder(args))
-                .Construct()
-                .Run();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); });
-    }
-}
+ {
+     public class Program
+     {
+         public static void Main(string[] args)
+         {
+             var appBuilder = WebApplication.CreateBuilder(args);
+             appBuilder.Services.AddControllers();
+ 
+             var app = appBuilder.UseFramework()
+                 .AddServices<ConsoleCommandsExecutingServicesModule>()
+                 .AddServices<ServerServicesModule>()
+                 .AddComponent<TerminalConsoleSubsystem>()
+                 .AddComponent<DebugStartupScript>()
+                 .Construct()
+                 .Asp();
+ 
+             app.MapControllers();
+             app.UseWebSockets()
+                 .UseHttpsRedirection();
+ 
+             app.Run();
+         }
+     }
+ }
